requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
psycopg2-binary
jinja2
pyyaml
cython
pyzabbix
seaborn
flask
plotly
plotly-express
fastdtw
